requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.5
pinecone==7.3.0
httpx==0.25.2
python-multipart==0.0.6
alembic==1.12.1
apscheduler==3.10.4

